Import of MIROD/XML-Twig-3.23 from CPAN.
gitpan-cpan-distribution: XML-Twig
gitpan-cpan-version:      3.23
gitpan-cpan-path:         MIROD/XML-Twig-3.23.tar.gz
gitpan-cpan-author:       MIROD
gitpan-cpan-maturity:     released

+++ t/test_new_features_3.22.t
@@ -1,7 +1,7 @@
 #!/usr/bin/perl -w
 use strict;
 
-# $Id: test_new_features_3.22.t,v 1.6 2005/10/14 16:14:48 mrodrigu Exp $
+# $Id: test_new_features_3.22.t,v 1.8 2005/12/08 10:47:43 mrodrigu Exp $
 use Carp;
 
 use FindBin qw($Bin);
@@ -11,8 +11,14 @@ use tools;
 use XML::Twig;
 
 my $DEBUG=0;
-print "1..14\n";
+print "1..20\n";
 
+         
+{ my $doc= q{<?xml version="1.0" ?>
+<!DOCTYPE doc [ <!ENTITY foo 'toto'>]>
+<doc>&foo;</doc>};
+  XML::Twig->new( keep_encoding => 1)->parse( $doc);
+}
 
 { # testing parse_html
  
@@ -30,7 +36,7 @@ print "1..14\n";
       
     }
   else
-    { skip( 3 => 'need XML::LibXML to test parse_html'); }
+    { skip( 3 => 'need  HTML::TreeBuilder to test parse_html'); }
 }
 
 { # testing _use
@@ -52,13 +58,23 @@ print "1..14\n";
   is( XML::Twig->nparse( twig_handlers => { doc => sub { $_->set_tag( 'foo'); } }, $doc_file)->sprint, '<foo></foo>', 'nparse file and option');
   unlink $doc_file;
 
-  $doc=q{<html><head><title>foo</title></head><body><p>toto</p></body></html>}; 
-  is( XML::Twig->nparse( $doc)->sprint, $doc, 'nparse well formed html string');
-  $doc_file="doc.html";
-  spit( $doc_file => $doc);
-  is( XML::Twig->nparse( $doc_file)->sprint, $doc, 'nparse well formed html file');
-  #is( XML::Twig->nparse( "file://$doc_file")->sprint, $doc, 'nparse well formed url');
-  unlink $doc_file;
+if( _use 'HTML::TreeBuilder')
+  {
+      $doc=q{<html><head><title>foo</title></head><body><p>toto</p></body></html>}; 
+      is( XML::Twig->nparse( $doc)->sprint, $doc, 'nparse well formed html string');
+      $doc_file="doc.html";
+      spit( $doc_file => $doc);
+      is( XML::Twig->nparse( $doc_file)->sprint, $doc, 'nparse well formed html file');
+      #is( XML::Twig->nparse( "file://$doc_file")->sprint, $doc, 'nparse well formed url');
+      unlink $doc_file;
+
+      XML::Twig::_disallow_use( 'HTML::TreeBuilder');
+      eval{ XML::Twig->new->parse_html( '<html/>'); };
+      matches( $@, "^cannot parse HTML: missing HTML::TreeBuilder", "parse_html without HTML::TreeBuilder");
+      XML::Twig::_allow_use( 'HTML::TreeBuilder');
+  }
+else
+  { skip( 3); }
 
   if( _use 'HTML::TreeBuilder')
     { $doc=q{<html><head><title>foo</title></head><body><p>toto<br>tata</p></body></html>}; 
@@ -71,4 +87,50 @@ print "1..14\n";
     }
   else
     { skip ( 1); }
+}
+
+{ 
+  my $file= "$Bin/test_new_features_3.22.html";
+  if( -f $file) 
+    { XML::Twig::_disallow_use( 'LWP::Simple');
+      eval { XML::Twig->nparse( "file://$file"); };
+      matches( $@, "^missing LWP::Simple", "nparse html url without LWP::Simple");
+      XML::Twig::_allow_use( 'LWP::Simple');
+      if( XML::Twig::_use( 'LWP::Simple') && XML::Twig::_use( 'HTML::TreeBuilder'))
+        { my $content= XML::Twig->nparse( "file://$file")->sprint;
+          (my $expected= slurp( $file))=~ s{(<(meta|br)[^>]*>)}{$1</$2>}g;
+          $expected=~s{<p>t3}{<p>t3</p>};
+          is( $content, $expected, "nparse url");
+        }
+      else
+        { skip( 1 => "cannot test html url parsing without LWP::Simple and HTML::TreeBuilder"); }
+      
+    }
+  else
+    { skip( 2 => "cannot find $file"); }
+}
+
+{ 
+  my $file= "$Bin/test_new_features_3.22.xml";
+  if( -f $file) 
+    { XML::Twig::_disallow_use( 'LWP');
+      eval { XML::Twig->nparse( "file://$file"); };
+      matches( $@, "^LWP not available", "nparse url without LWP");
+      XML::Twig::_allow_use( 'LWP');
+      if( XML::Twig::_use( 'LWP'))
+        { my $content= XML::Twig->nparse( "file://$file")->sprint;
+          is( $content, "<doc></doc>", "nparse url");
+        }
+      else
+        { skip( 1 => "cannot test url parsing without LWP"); }
+    }
+  else
+    { skip( 2 => "cannot find $file"); }
+}
+ 
+
+{ my $file= "t/test_new_features_3.22.xml";
+  open( FH, "<$file") or die "cannot find test file '$file': $!";
+  my $content= XML::Twig->nparse( \*FH)->sprint;
+  is( $content, "<doc></doc>", "nparse glob");
 }
